Add group modal for viewing/editing a group
Still a lot of ui polishing to do. This has to be connected to API with
Queries ( update description, members, delete ) and device ( update members and delete )
but the basic structure is here.

You can view a group and directly edit the members ( only if the group
is APPROVED ).

NEXT STEP: create group flow.