use gate devices name in u2f_authenticate
It's a bit hacky but there is a mismatch between the username registered
on the gate, and the username that has been used to generate the
key_handle while registering on the device.

The gate instanciates the devices with fixed username like
`user = User('op10')` It cannot be changed on the fly. So we have to
maintain a mapping between device number and name used by the gate.