in transaction status, display 'NB+' confirmation
At some point, the gate won't fetch the number of confirmations anymore
because it is not relevant anymore and it save some expensive network calls.

Now the gate returns the threshold as `min_confirmations` field in
transaction. If the number of confirmation is >= to this number we
display `${min_confirmations}+` ( for example 30+ )