enrich data with a type field
Every time we are doing a query or a connectionQuery we take the schema
key from the object and enrich the result with it, so we can use it with
flowtype.
It's a at the same level of the transformations utils we already do for
transforming strings number to BigNumber objects. Each query can already
define a `transformer` method. So we just compose a potential
transformer defined with a one we add the type field