disable re-create button with tooltip for the reason
Instead of always hiding the re-create button when it's not possible to
replay a request, we show it disable with a tooltip explaining the
reason why it's not possible. For some cases we hide it anyway because
it does not make sense to disable it. For example if user is an admin
and is looking at a transaction.